requirements.txt
gitpython>=3.1.40
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pathspec>=0.11.2
